// src/app/models/task-technician.ts
export class TaskTechnician{
    TaskTechnicianID: number;
    JobCardID: number;
    JobTaskID: number;
    EmployeeID: number;
    ActualTimeSpent: Date;
    IsAccepted: boolean;
    AcceptedDateTime: Date;
    EndDateTime: Date;
    IsCompleted: boolean;
    IsDeleted: boolean;
    CreatedBy: number;
    ModifiedBy: number;
    Day:string;
    TaskCount:number;
    TaskDescription:string;
    PlateNumber:string;
    TotalTimeSpent:number;
    TotalCost:number;
    TotalTasks:number;
    IsModified: number;
    WorkshopID: number;
}